New input method
Added a new input method using a virtual driver (run as administrator)
Added a setting to suppress the build-up of Recoil during prolonged firing.
Added the ability to turn off the sound to turn on/off the macro.
Added the ability to hide the macro window(F6 hotkey)

FIXED:
When saving changes to an existing preset, the changed preset name was not displayed, it was displayed only when the preset was already selected, but remained unchanged in the list.